Re-organize tile server structure
Closes #2304

Does a few things that are related, but somewhat distinct
 - Ensures that only one database pool is instantiated
 - Organizes server similar to API server organization
 - Removes circular dependencies between Stitch, GlobalSummary,
 Mosaic, TileSources, and LayerCache
 - Deletes unused and unnecessary files/functions
 - Adds docstrings to most methods
 - Consolidates some logic
 - Removes most implicit arguments to function calls
