+++ app-backend/tile/src/main/scala/Router.scala
@@ -1,55 +1,47 @@
 package com.azavea.rf.tile
 
+import com.azavea.rf.common.cache.CacheClient
+import com.azavea.rf.common.cache.kryo.KryoMemcachedClient
 import com.azavea.rf.database.Database
-import com.azavea.rf.tile.routes._
-import com.azavea.rf.tile.tool._
+import com.azavea.rf.tile.healthcheck.HealthCheckRoute
+import com.azavea.rf.tile.projects.ProjectRoutes
+import com.azavea.rf.tile.tools.ToolRoutes
 
 import ch.megard.akka.http.cors.CorsDirectives._
 import ch.megard.akka.http.cors.CorsSettings
-import akka.http.scaladsl.model.{HttpResponse, StatusCodes}
-import akka.http.scaladsl.server._
 import com.typesafe.scalalogging.LazyLogging
+import geotrellis.spark.io.postgres.PostgresAttributeStore
 
-class Router extends LazyLogging
-    with TileAuthentication
-    with TileErrorHandler {
+class Router extends HealthCheckRoute
+  with ProjectRoutes
+  with ToolRoutes
+  with LazyLogging
+  with TileAuthentication
+  with TileErrorHandler {
 
   implicit lazy val database = Database.DEFAULT
   val system = AkkaSystem.system
   implicit val materializer = AkkaSystem.materializer
 
-  lazy val blockingSceneRoutesDispatcher =
+  implicit lazy val blockingDispatcher =
     system.dispatchers.lookup("blocking-dispatcher")
+  val memcachedClient: KryoMemcachedClient = KryoMemcachedClient.DEFAULT
 
-  val toolRoutes = new ToolRoutes()
+  val rfCache: CacheClient = new CacheClient(memcachedClient)
+  val store = PostgresAttributeStore()
 
   val corsSettings = CorsSettings.defaultSettings
 
   def root = cors() {
     handleExceptions(tileExceptionHandler) {
       pathPrefix("tiles") {
-        pathPrefix(JavaUUID) { projectId =>
-          tileAccessAuthorized(projectId) {
-            case true => MosaicRoutes.mosaicProject(projectId)(database)
-            case _ => reject(AuthorizationFailedRejection)
-          }
-        } ~
-        pathPrefix("healthcheck") {
+        projectRoutes ~ pathPrefix("healthcheck") {
           pathEndOrSingleSlash {
-            get {
-              HealthCheckRoute.root
-            }
+            healthCheckRoute
           }
         } ~
         pathPrefix("tools") {
-          get {
-            tileAuthenticateOption { _ =>
-              toolRoutes.tms(TileSources.cachedTmsSource) ~
-              toolRoutes.validate ~
-              toolRoutes.histogram ~
-              toolRoutes.preflight
-            }
-          }
+          toolRoutes
         }
       }
     }
