--- app-backend/tile/src/main/scala/routes/HealthCheckRoute.scala
@@ -1,16 +1,0 @@
-package com.azavea.rf.tile.routes
-
-import akka.http.scaladsl.server._
-import akka.http.scaladsl.server.Directives._
-import com.typesafe.scalalogging.LazyLogging
-import de.heikoseeberger.akkahttpcirce.CirceSupport._
-
-object HealthCheckRoute extends LazyLogging {
-  def root: Route = complete {
-    Map(
-      "service" -> "tile",
-      "status" -> "OK",
-      "active threads" -> Thread.activeCount.toString
-    )
-  }
-}
